+++ buildSrc/src/main/java/Plugins.kt
@@ -1,4 +1,5 @@
 object Plugins {
+    const val compose = "org.jetbrains.kotlin.plugin.compose"
     const val hilt = "com.google.dagger.hilt.android"
     const val ksp = "com.google.devtools.ksp"
 }